+++ CsharpExtras/Enumerable/NonEmpty/INonEmptyCollection.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CsharpExtras.Enumerable.NonEmpty
+namespace CsharpExtras._Enumerable.NonEmpty
 {
     public interface INonEmptyCollection<T> : ICollection<T>
     {